レシピ機能の整理 - 不要なレシピ関連のコンポーネントを削除し、レイアウトファイル内のNavbarコンポーネントをコメントアウトしました。また、MainRecipesページの構造を更新し、ユーザーヘッダーとレシピリストを表示するようにしました。
