+++ my-app/app/user/recipes/page.tsx
@@ -1,38 +1,23 @@
-import { Suspense } from "react"
-import RecipeGrid from "@/components/recipe-grid"
-import RecipeFilters from "@/components/recipe-filters"
-import AiRecipeSuggestion from "@/components/ai-recipe-suggestion"
-import { Button } from "@/components/ui/button"
-import { PlusCircle, Loader2 } from "lucide-react"
-import Link from "next/link"
+import RecipesLists from "@/components/Recipes/RecipeLists";
+import UserHeader from "@/components/Recipes/user-header";
+import { Suspense } from "react";
 
-export default function RecipePage() {
+export default async function MainRecipes() {
   return (
-    <main className="container mx-auto px-4 pb-24 pt-4">
-      <div className="flex justify-between items-center mb-6">
-        <h1 className="text-2xl font-bold">Recipes</h1>
-        <Link href="/user/recipes/create">
-          <Button size="sm" className="flex items-center gap-1">
-            <PlusCircle className="h-4 w-4" />
-            <span>New Recipe</span>
-          </Button>
-        </Link>
-      </div>
-
-      <AiRecipeSuggestion />
-
-      <RecipeFilters />
-
+    <div className="flex flex-col items-center justify-center min-h-screen">
       <Suspense
         fallback={
-          <div className="flex justify-center items-center h-64">
-            <Loader2 className="h-8 w-8 animate-spin text-primary" />
-            <span className="sr-only">Loading</span>
+          <div className="flex flex-col justify-center items-center h-screen w-screen absolute top-0 left-0">
+            <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-gray-900"></div>
+            <p className="text-gray-500">Loading...</p>
           </div>
         }
       >
-        <RecipeGrid />
+        <div className="flex flex-col items-center">
+          <UserHeader />
+          <RecipesLists />
+        </div>
       </Suspense>
-    </main>
-  )
+    </div>
+  );
 }
